Cambiar enteros por numeros de punto flotante

+++ CFunctions/com_baiper06_jni_Main.c
@@ -10,20 +10,20 @@
 #include "../JavaCalc/bin/com_baiper06_jni_Main.h"
 #include <stdio.h>
 
-JNIEXPORT int JNICALL Java_com_baiper06_jni_Main_add(JNIEnv *env, jobject obj, int pNumberX, int pNumberY){
+JNIEXPORT float JNICALL Java_com_baiper06_jni_Main_add(JNIEnv *env, jobject obj, float pNumberX, float pNumberY){
 	return pNumberX+pNumberY;
 }
 
-JNIEXPORT int JNICALL Java_com_baiper06_jni_Main_subtraction(JNIEnv *env, jobject obj, int pNumberX, int pNumberY){
+JNIEXPORT float JNICALL Java_com_baiper06_jni_Main_subtraction(JNIEnv *env, jobject obj, float pNumberX, float pNumberY){
 	return pNumberX-pNumberY;
 }
 
 
-JNIEXPORT int JNICALL Java_com_baiper06_jni_Main_division(JNIEnv *env, jobject obj, int pNumberX, int pNumberY){
+JNIEXPORT float JNICALL Java_com_baiper06_jni_Main_division(JNIEnv *env, jobject obj, float pNumberX, float pNumberY){
 	return pNumberX/pNumberY;
 }
 
 
-JNIEXPORT int JNICALL Java_com_baiper06_jni_Main_multiplication(JNIEnv *env, jobject obj, int pNumberX, int pNumberY){
+JNIEXPORT float JNICALL Java_com_baiper06_jni_Main_multiplication(JNIEnv *env, jobject obj, float pNumberX, float pNumberY){
 	return pNumberX*pNumberY;
 }
